refactor(orchestrator): extract drt validation into its own module
